fix(process-details): avoid double-click on delete button (#365)
* Disable delete button when server call is incoming. This allow to avoid accidental double click on delete button triggering delete call twice and leading to some unwanted logs.
* Do not display error when closing delete pop up by clicking on backdrop

Relates to [BS-16690](https://bonitasoft.atlassian.net/browse/BS-16690)

+++ main/features/admin/processes/details/delete-process-modal.controller.js
@@ -8,13 +8,22 @@
   function DeleteProcessModalInstanceCtrl(processAPI, process, $modalInstance) {
     var vm = this;
     vm.process = process;
+    vm.processing = false;
 
     vm.delete = function () {
+      // avoid double click
+      if (vm.processing) {
+        return;
+      }
+
+      vm.processing = true;
       processAPI.delete({id: process.id}).$promise
         .then(function () {
           $modalInstance.close();
+          vm.processing = false;
         }, function (error) {
           $modalInstance.dismiss(error);
+          vm.processing = false;
         });
     };
